+++ playground/spa/src/device.ts
@@ -1,15 +1,15 @@
 export const DEVICE = [
-  { isDevice: (clientWidth: number) => clientWidth <= 750, UIWidth: 375, deviceWidthRange: [300, 750], type: '手机' }, // 手机
+  { match: (clientWidth: number) => clientWidth <= 750, base: 375, range: [300, 750], type: '手机' }, // 手机
   {
-    isDevice: (clientWidth: number) => clientWidth > 750 && clientWidth <= 1280,
-    UIWidth: 1280,
-    deviceWidthRange: [960, 1280],
+    match: (clientWidth: number) => clientWidth > 750 && clientWidth <= 1280,
+    base: 1280,
+    range: [960, 1280],
     type: '平板',
   }, // 平板
   {
-    isDevice: (clientWidth: number) => clientWidth > 1280,
-    UIWidth: 1920,
-    deviceWidthRange: [1280, 1920],
+    match: (clientWidth: number) => clientWidth > 1280,
+    base: 1920,
+    range: [1280, 1920],
     type: '电脑',
   }, // 电脑
 ]