+++ playground/spa/src/flexible.ts
@@ -1,0 +1,4 @@
+import { flexible } from 'modern-flexible'
+import { DEVICE } from './device'
+
+flexible({ devices: DEVICE, rootValue: 16 })