+++ playground/spa/vite.config.ts
@@ -1,20 +1,8 @@
 import react from '@vitejs/plugin-react'
 import { defineConfig } from 'vite'
-import { injectScripts, publicTypescript } from 'vite-plugin-public-typescript'
 
 // https://vitejs.dev/config/
 export default defineConfig(() => ({
   base: '/modern-flexible/',
-  plugins: [
-    react(),
-    publicTypescript(),
-    injectScripts((manifest) => [
-      {
-        attrs: {
-          src: manifest.flexible,
-        },
-        injectTo: 'head',
-      },
-    ]),
-  ],
+  plugins: [react()],
 }))